refactor: reorganize time picker components and enhance date formatting utilities
- Removed the old TimePicker component and replaced it with a more modular structure.
- Updated the import path for TimePicker in ConcurrentChartClient for better clarity.
- Added new date formatting functions to improve readability and usability in date/time components.
- Cleaned up the .gitignore file by removing unnecessary entries.
